fix: minor bug fix in erase
